WIP: line-request: don't accept NULL line config
While it was initially agreed that NULL line_config passed to
gpiod_chip_request_lines() would result in lines being requested with
default config (just like a newly created struct gpiod_line_config),
I believe this is unintuitive in the context of all other functions in
the C library not being NULL-aware (as advertised in one of the first
paragraphs of the documentation) and should be changed to make both
config structures mandatory.

This is C after all and it should be both explicit and spartan. :)

Signed-off-by: Bartosz Golaszewski <brgl@bgdev.pl>
Reviewed-by: Kent Gibson <warthog618@gmail.com>