WIP: API: add an enum for line values
In order to explicitly stress that line values as understood by libgpiod
are logical values, expose a two-value enum with values called: ACTIVE
and INACTIVE that should be used whenever referring to the state of GPIO
lines.

The value of INACTIVE is set to 0 while that of ACTIVE to 1 so that users
can still use integers in C (where no scoped enums exist).

Signed-off-by: Bartosz Golaszewski <brgl@bgdev.pl>