WIP: gpiosim: remove randomness from configfs name generation
None of the test suites use the random name generation at this point so
it's very much dead code and also every test-suite implements its own
deterministic way of assigning file names to configfs attributes while
also doing it wrong (as is evident as soon as we try to remove one
simulator but not all of them and then create another one).

This removes the 'name' argument from gpiosim_dev_new() and
gpiosim_bank_new() and makes the generation of the file names for
configfs attributes deterministic at the libgpiosim level. The names
are created from the program invocation name, the PID and an ID allocated
using a binary search tree in a similar manner to kernel's IDA API (find
the lowest available integer for a new name).

The ID pool is global within the library code and as such has protection
using a mutex (unlike the context object and its children which are
expected to be protected by the user if needed).

Signed-off-by: Bartosz Golaszewski <brgl@bgdev.pl>